split WC and MC login functionality clearly (#699)
